# speed up multiple invocations of counterfeiter