--- a/fixtures/aliased_interfaces.go
+++ b/fixtures/aliased_interfaces.go
@@ -2,7 +2,7 @@ package fixtures
 
 import alias "github.com/maxbrunsfeld/counterfeiter/v6/fixtures/another_package"
 
-//go:generate counterfeiter . AliasedInterface
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . AliasedInterface
 
 // AliasedInterface is an interface that embeds an interface in an aliased package.
 type AliasedInterface interface {