--- a/fixtures/aliased_package/in_aliased_package.go
+++ b/fixtures/aliased_package/in_aliased_package.go
@@ -1,6 +1,6 @@
 package the_aliased_package // import "github.com/maxbrunsfeld/counterfeiter/v6/fixtures/aliased_package"
 
-//go:generate counterfeiter . InAliasedPackage
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . InAliasedPackage
 type InAliasedPackage interface {
 	Stuff(int) string
 }