--- a/fixtures/another_package/types.go
+++ b/fixtures/another_package/types.go
@@ -2,7 +2,7 @@ package another_package // import "github.com/maxbrunsfeld/counterfeiter/v6/fixt
 
 type SomeType int
 
-//go:generate counterfeiter . AnotherInterface
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . AnotherInterface
 type AnotherInterface interface {
 	AnotherMethod([]SomeType, map[SomeType]SomeType, *SomeType, SomeType, chan SomeType)
 }