--- a/fixtures/compound_return.go
+++ b/fixtures/compound_return.go
@@ -1,6 +1,6 @@
 package fixtures
 
-//go:generate counterfeiter . SomethingElse
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . SomethingElse
 type SomethingElse interface {
 	ReturnStuff() (a, b int)
 }