--- a/fixtures/dup_packages/a/a.go
+++ b/fixtures/dup_packages/a/a.go
@@ -2,7 +2,7 @@ package a // import "github.com/maxbrunsfeld/counterfeiter/v6/fixtures/dup_packa
 
 import "github.com/maxbrunsfeld/counterfeiter/v6/fixtures/dup_packages/a/foo"
 
-//go:generate counterfeiter . A
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . A
 type A interface {
 	V1() foo.I
 }