--- a/fixtures/dup_packages/a/foo/a.go
+++ b/fixtures/dup_packages/a/foo/a.go
@@ -2,7 +2,7 @@ package foo // import "github.com/maxbrunsfeld/counterfeiter/v6/fixtures/dup_pac
 
 type S struct{}
 
-//go:generate counterfeiter . I
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . I
 type I interface {
 	FromA() S
 }