--- a/fixtures/dup_packages/dupAB.go
+++ b/fixtures/dup_packages/dupAB.go
@@ -1,6 +1,6 @@
 package dup_packages // import "github.com/maxbrunsfeld/counterfeiter/v6/fixtures/dup_packages"
 
-//go:generate counterfeiter . DupAB
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . DupAB
 type DupAB interface {
 	DupA
 	DupB