--- a/fixtures/dup_packages/dupB.go
+++ b/fixtures/dup_packages/dupB.go
@@ -2,7 +2,7 @@ package dup_packages // import "github.com/maxbrunsfeld/counterfeiter/v6/fixture
 
 import "github.com/maxbrunsfeld/counterfeiter/v6/fixtures/dup_packages/b/foo"
 
-//go:generate counterfeiter . DupB
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . DupB
 type DupB interface {
 	B() foo.S
 }