--- a/fixtures/dup_packages/foo/multi_import.go
+++ b/fixtures/dup_packages/foo/multi_import.go
@@ -7,7 +7,7 @@ import (
 
 type S struct{}
 
-//go:generate counterfeiter . MultiAB
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . MultiAB
 type MultiAB interface {
 	Mine() S
 	foo.I