--- a/fixtures/embeds_interfaces.go
+++ b/fixtures/embeds_interfaces.go
@@ -6,7 +6,7 @@ import (
 	"github.com/maxbrunsfeld/counterfeiter/v6/fixtures/another_package"
 )
 
-//go:generate counterfeiter . EmbedsInterfaces
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . EmbedsInterfaces
 type EmbedsInterfaces interface {
 	http.Handler
 	another_package.AnotherInterface