--- a/fixtures/has_other_types.go
+++ b/fixtures/has_other_types.go
@@ -1,6 +1,6 @@
 package fixtures
 
-//go:generate counterfeiter . HasOtherTypes
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . HasOtherTypes
 type HasOtherTypes interface {
 	GetThing(SomeString) SomeFunc
 }