--- a/fixtures/has_var_args.go
+++ b/fixtures/has_var_args.go
@@ -1,12 +1,12 @@
 package fixtures
 
-//go:generate counterfeiter . HasVarArgs
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . HasVarArgs
 type HasVarArgs interface {
 	DoThings(int, ...string) int
 	DoMoreThings(int, int, ...string) int
 }
 
-//go:generate counterfeiter . HasVarArgsWithLocalTypes
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . HasVarArgsWithLocalTypes
 type HasVarArgsWithLocalTypes interface {
 	DoThings(...LocalType)
 }