--- a/fixtures/hyphenated_package_same_name/some_package/interface.go
+++ b/fixtures/hyphenated_package_same_name/some_package/interface.go
@@ -2,7 +2,7 @@ package some_package // import "github.com/maxbrunsfeld/counterfeiter/v6/fixture
 
 import "github.com/maxbrunsfeld/counterfeiter/v6/fixtures/hyphenated_package_same_name/hyphen-ated/some_package"
 
-//go:generate counterfeiter . SomeInterface
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . SomeInterface
 type SomeInterface interface {
 	CreateThing() some_package.Thing
 }