--- a/fixtures/imports_go_hyphen_package.go
+++ b/fixtures/imports_go_hyphen_package.go
@@ -4,7 +4,7 @@ import (
 	"github.com/maxbrunsfeld/counterfeiter/v6/fixtures/go-hyphenpackage"
 )
 
-//go:generate counterfeiter . ImportsGoHyphenPackage
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . ImportsGoHyphenPackage
 type ImportsGoHyphenPackage interface {
 	UseHyphenType(hyphenpackage.HyphenType)
 }