--- a/fixtures/reuses_arg_types.go
+++ b/fixtures/reuses_arg_types.go
@@ -1,6 +1,6 @@
 package fixtures
 
-//go:generate counterfeiter . ReusesArgTypes
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . ReusesArgTypes
 type ReusesArgTypes interface {
 	DoThings(x, y string)
 }