--- a/fixtures/something.go
+++ b/fixtures/something.go
@@ -1,6 +1,6 @@
 package fixtures
 
-//go:generate counterfeiter . Something
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . Something
 type Something interface {
 	DoThings(string, uint64) (int, error)
 	DoNothing()