--- a/fixtures/something_remote.go
+++ b/fixtures/something_remote.go
@@ -2,7 +2,7 @@ package fixtures
 
 import the_aliased_package "github.com/maxbrunsfeld/counterfeiter/v6/fixtures/aliased_package"
 
-//go:generate counterfeiter . SomethingWithForeignInterface
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . SomethingWithForeignInterface
 
 // SomethingWithForeignInterface is an interface that embeds a foreign interface.
 type SomethingWithForeignInterface interface {