--- a/fixtures/sync/interface.go
+++ b/fixtures/sync/interface.go
@@ -1,6 +1,6 @@
 package sync // import "github.com/maxbrunsfeld/counterfeiter/v6/fixtures/sync"
 
-//go:generate counterfeiter . SyncSomething
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . SyncSomething
 type SyncSomething interface {
 	DoThings(string, uint64) (int, error)
 	DoNothing()