--- a/fixtures/typed_function.go
+++ b/fixtures/typed_function.go
@@ -1,4 +1,4 @@
 package fixtures
 
-//go:generate counterfeiter . SomethingFactory
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . SomethingFactory
 type SomethingFactory func(string, map[string]interface{}) string