--- a/fixtures/unexported.go
+++ b/fixtures/unexported.go
@@ -1,9 +1,9 @@
 package fixtures
 
-//go:generate counterfeiter . unexportedFunc
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . unexportedFunc
 type unexportedFunc func(string, map[string]interface{}) string
 
-//go:generate counterfeiter . unexportedInterface
+//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 . unexportedInterface
 type unexportedInterface interface {
 	Method(string, map[string]interface{}) string
 }