--- a/generator/cache.go
+++ b/generator/cache.go
@@ -1,0 +1,33 @@
+package generator
+
+import "golang.org/x/tools/go/packages"
+
+type Cache struct {
+	packageMap map[string]interface{}
+}
+
+type FakeCache struct{}
+
+func (c *FakeCache) Load(workingDir string) ([]*packages.Package, bool)    { return nil, false }
+func (c *FakeCache) Store(workingDir string, packages []*packages.Package) {}
+
+type Cacher interface {
+	Load(workingDir string) ([]*packages.Package, bool)
+	Store(workingDir string, packages []*packages.Package)
+}
+
+func (c *Cache) Load(workingDir string) ([]*packages.Package, bool) {
+	p, ok := c.packageMap[workingDir]
+	if !ok {
+		return nil, false
+	}
+	packages, ok := p.([]*packages.Package)
+	return packages, ok
+}
+
+func (c *Cache) Store(workingDir string, packages []*packages.Package) {
+	if c.packageMap == nil {
+		c.packageMap = map[string]interface{}{}
+	}
+	c.packageMap[workingDir] = packages
+}