--- a/generator/fake.go
+++ b/generator/fake.go
@@ -48,7 +48,7 @@ type Method struct {
 
 // NewFake returns a Fake that loads the package and finds the interface or the
 // function.
-func NewFake(fakeMode FakeMode, targetName string, packagePath string, fakeName string, destinationPackage string, workingDir string) (*Fake, error) {
+func NewFake(fakeMode FakeMode, targetName string, packagePath string, fakeName string, destinationPackage string, workingDir string, cache Cacher) (*Fake, error) {
 	f := &Fake{
 		TargetName:         targetName,
 		TargetPackage:      packagePath,
@@ -59,7 +59,7 @@ func NewFake(fakeMode FakeMode, targetName string, packagePath string, fakeName 
 	}
 
 	f.Imports.Add("sync", "sync")
-	err := f.loadPackages(workingDir)
+	err := f.loadPackages(cache, workingDir)
 	if err != nil {
 		return nil, err
 	}