… .circleci/config.yml
… .travis.yml
… arguments/flags.go
… arguments/parser_test.go
… arguments/parser_windows_test.go
… benchmark_test.go
… command/runner_test.go
… generator/generator_internals_test.go
… go.mod
… go.sum
… integration/roundtrip_test.go
… main.go
… scripts/ci.sh